+++ plugins/frontend/backstage-plugin-github-pull-requests/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @roadiehq/backstage-plugin-github-pull-requests
 
+## 3.1.0
+
+### Minor Changes
+
+- 5dba102: Create a shareable package to manage GitHub plugins auth state. This allows plugins using GitHub OAuth to use common functionality to show/hide components based on logged in scopes and urls
+
+### Patch Changes
+
+- Updated dependencies [5dba102]
+  - @roadiehq/github-auth-utils-react@0.2.0
+
 ## 3.0.1
 
 ### Patch Changes